Bring up for tiny4412 board.
1. I refered to samsung origen board configurations.
2. Add the compile target in Makefile to generate u-boot-spl.dis.
	usage:
	make u-boot-spl.bin
	or
	make spl/u-boot-spl.bin
3. At present, I only solved compile problems. And the build image
can not be used yet.
4. At the begining, I want to refer samsung odroid or trats2, but I
found that those two boards's configurations were not suitable for
tiny4412, for more information, can refor to the pdf file:

	Android_Exynos4412_iROM_Secure_Booting_Guide_Ver.1.00.00.pdf

In those two boards, there is no spl image (can be used as BL2), and
the relative address of function board_init_f is bigger than 14k, we
all know that the size of BL2 image can not be bigger than 14k.