Reach to uboot.
Although can reach to u-boot, this is not the best.
The original code is to copy u-boot from sdcard to dram directly.But
there maybe some problems that can't do this. So I take a around way.

I use iram (0x02050000 - 0x02060000) as an buffer, firstly I copy
u-boot from sdcard to this buffer, then copy it to dram.

I will consider this issue in some other days.